source: only do tracing on speaker, fix y/z-offset
Previously, the speaker acoustic center offsets were mistakenly
applied to the X and Y axes. In reality, there can never be an offset
in the X direction, since we treat X as the direction in which the
speaker points and the acoustic center must always be on the face of the
loudspeaker. This commit fixes that mistake.

The only sources we do tracing on are speakers, so make that method only
apply to speakers. This avoids the awkward behaviour where both the
LoudspeakerSpec **and** the vanilla Source struct defined a dispersion,
only one of which would be used. This is a cleaner solution
unless/until we need to support another source type other than a
speaker. For now, we don't know of any, so we won't worry about it.