listening_triangle: fix sign of Gervais constant
The listening position should be slightly INSIDE the equilateral
triangle defined by the loudspeakers so that the path from the speakers
intersects the engineer's ears, not the center of his/her head. Rod
Gervais calculates this distance as 0.38m.

Previously, the code contained a bug that ADDED this constant to the
distance from the reference point, rather than subtracting it. This
commit fixes that mistake, with the effect of moving the listening
position 76cm forward in all simulations. This is the correct behavior.