Add trace mode
Trace mode allows specifying a particular point through with to fire one
or more shots. If num-shots is greater than 1, shots are sampled across
a configurable angular spread.

In trace mode, all reflections are captured out to the specified min
gain or max ITD and stored in annotations.json. This allows visualizing
the path followed by all the reflections for the specified shot(s),
regardless of whether they ever arrive back at the listening position.

This is especially useful for visualizing how some feature of the room
prevented a shot from causing a problematic reflection that you might
have expected to.