Replace fogleman/pt with our own version
Previously, we were keeping a hacked version of fogleman/pt in vendor.
This was annoying and constantly caused vendoring issues. Since we
hacked it specifically for this project, it makes sense to keep our
modified version in the project instead of in vendor.

This commit moves our hacked version out of vendor and fixes our
dependencies to point to that version instead fogleman/pt.